refactor: move chart types into separate file from utils
