+++ src/explore-education-statistics-common/src/modules/find-statistics/components/charts/__tests__/VerticalBarBlock.test.tsx
@@ -1,9 +1,9 @@
 import {
-  AxesConfiguration,
   ChartMetaData,
   ChartProps,
-} from '@common/modules/find-statistics/components/charts/util/chartUtils';
+} from '@common/modules/find-statistics/components/charts/types/chart';
 import { DataBlockData } from '@common/services/dataBlockService';
+import { AxesConfiguration } from '@common/services/publicationService';
 import React from 'react';
 
 import { render } from 'react-testing-library';
